Fix: cleaned out code for TCP und TCPv6
*some differences might not be necessary
*SIGINT will stop process on first attempt
